split the logic into components, moved the helper function to a separate file
