+++ src/App.tsx
@@ -1,12 +1,20 @@
 /* eslint-disable jsx-a11y/label-has-associated-control */
-import React, { useEffect, useState } from 'react';
-import { UserWarning } from './UserWarning';
-import { getTodos, USER_ID } from './api/todos';
-import { getFilteredTodos, Todo } from './types/Todo';
+import { FC, useEffect, useState } from 'react';
 import classNames from 'classnames';
-import { TodoStatusFilter } from './types/TodoStatusFilter';
 
-export const App: React.FC = () => {
+import { Todo } from './types/Todo';
+import { UserWarning } from './UserWarning';
+
+import { TodoStatusFilter } from './types/TodoStatusFilter';
+import { getTodos, USER_ID } from './api/todos';
+import { getFilteredTodos } from './utils/filterTodos';
+
+import { TodoItem } from './components/TodoItem';
+import { TodoForm } from './components/TodoForm';
+import { TodoFooter } from './components/TodoFooter';
+import { ErrorNotification } from './components/ErrorNotification';
+
+export const App: FC = () => {
   const [todos, setTodos] = useState<Todo[]>([]);
   const [errorMessage, setErrorMessage] = useState('');
   const [selectedFilter, setSelectedFilter] = useState(TodoStatusFilter.ALL);
@@ -26,7 +34,6 @@ export const App: React.FC = () => {
   }, [errorMessage]);
 
   useEffect(() => {
-    setErrorMessage('');
     getTodos()
       .then(setTodos)
       .catch(() => setErrorMessage('Unable to load todos'));
@@ -37,6 +44,14 @@ export const App: React.FC = () => {
   if (!USER_ID) {
     return <UserWarning />;
   }
+
+  const handleChangeStatus = (status: TodoStatusFilter) => {
+    setSelectedFilter(status);
+  };
+
+  const handleCloseErrorMessage = () => {
+    setErrorMessage('');
+  };
 
   return (
     <div className="todoapp">
@@ -52,148 +67,30 @@ export const App: React.FC = () => {
             data-cy="ToggleAllButton"
           />
 
-          {/* Add a todo on form submit */}
-          <form>
-            <input
-              data-cy="NewTodoField"
-              type="text"
-              className="todoapp__new-todo"
-              placeholder="What needs to be done?"
-            />
-          </form>
+          <TodoForm />
         </header>
 
         {filteredTodos.length !== 0 && (
           <section className="todoapp__main" data-cy="TodoList">
             {filteredTodos.map(todo => (
-              <div
-                key={todo.id}
-                data-cy="Todo"
-                className={classNames('todo', { completed: todo.completed })}
-              >
-                <label className="todo__status-label">
-                  <input
-                    data-cy="TodoStatus"
-                    type="checkbox"
-                    className="todo__status"
-                    defaultChecked={todo.completed}
-                  />
-                </label>
-
-                <span data-cy="TodoTitle" className="todo__title">
-                  {todo.title}
-                </span>
-
-                <button
-                  type="button"
-                  className="todo__remove"
-                  data-cy="TodoDelete"
-                >
-                  ×
-                </button>
-
-                {/* This form is shown instead of the title and remove button */}
-                {/* <form>
-              <input
-                data-cy="TodoTitleField"
-                type="text"
-                className="todo__title-field"
-                placeholder="Empty todo will be deleted"
-                value="Todo is being edited now"
-              />
-            </form> */}
-
-                {/* overlay will cover the todo while it is being deleted or updated */}
-                {/* 'is-active' class puts this modal on top of the todo */}
-                <div data-cy="TodoLoader" className="modal overlay">
-                  <div className="modal-background has-background-white-ter" />
-                  <div className="loader" />
-                </div>
-              </div>
+              <TodoItem key={todo.id} todo={todo} />
             ))}
           </section>
         )}
 
         {todos.length !== 0 && (
-          <footer className="todoapp__footer" data-cy="Footer">
-            <span className="todo-count" data-cy="TodosCounter">
-              {todos.filter(todo => !todo.completed).length} items left
-            </span>
-
-            <nav className="filter" data-cy="Filter">
-              <a
-                href="#/"
-                className={classNames('filter__link', {
-                  selected: selectedFilter === TodoStatusFilter.ALL,
-                })}
-                data-cy="FilterLinkAll"
-                onClick={() => setSelectedFilter(TodoStatusFilter.ALL)}
-              >
-                All
-              </a>
-
-              <a
-                href="#/active"
-                className={classNames('filter__link', {
-                  selected: selectedFilter === TodoStatusFilter.ACTIVE,
-                })}
-                data-cy="FilterLinkActive"
-                onClick={() => setSelectedFilter(TodoStatusFilter.ACTIVE)}
-              >
-                Active
-              </a>
-
-              <a
-                href="#/completed"
-                className={classNames('filter__link', {
-                  selected: selectedFilter === TodoStatusFilter.COMPLETED,
-                })}
-                data-cy="FilterLinkCompleted"
-                onClick={() => setSelectedFilter(TodoStatusFilter.COMPLETED)}
-              >
-                Completed
-              </a>
-            </nav>
-
-            <button
-              type="button"
-              className="todoapp__clear-completed"
-              data-cy="ClearCompletedButton"
-              disabled={!todos.some(todo => todo.completed)}
-            >
-              Clear completed
-            </button>
-          </footer>
+          <TodoFooter
+            todos={todos}
+            selectedFilter={selectedFilter}
+            onStatusChange={handleChangeStatus}
+          />
         )}
       </div>
 
-      <div
-        data-cy="ErrorNotification"
-        className={classNames(
-          'notification',
-          'is-danger',
-          'is-light',
-          'has-text-weight-normal',
-          { hidden: !errorMessage.length },
-        )}
-      >
-        <button
-          data-cy="HideErrorButton"
-          type="button"
-          className="delete"
-          onClick={() => setErrorMessage('')}
-        />
-        {errorMessage}
-        {/* Unable to load todos
-        <br />
-        Title should not be empty
-        <br />
-        Unable to add a todo
-        <br />
-        Unable to delete a todo
-        <br />
-        Unable to update a todo */}
-      </div>
+      <ErrorNotification
+        errorMessage={errorMessage}
+        onCloseErrorMessage={handleCloseErrorMessage}
+      />
     </div>
   );
 };
