+++ src/components/ErrorNotification.tsx
@@ -1,0 +1,42 @@
+import classNames from 'classnames';
+import { FC } from 'react';
+
+type Props = {
+  errorMessage: string;
+  onCloseErrorMessage: () => void;
+};
+
+export const ErrorNotification: FC<Props> = ({
+  errorMessage,
+  onCloseErrorMessage,
+}) => {
+  return (
+    <div
+      data-cy="ErrorNotification"
+      className={classNames(
+        'notification',
+        'is-danger',
+        'is-light',
+        'has-text-weight-normal',
+        { hidden: !errorMessage.length },
+      )}
+    >
+      <button
+        data-cy="HideErrorButton"
+        type="button"
+        className="delete"
+        onClick={onCloseErrorMessage}
+      />
+      {errorMessage}
+      {/* Unable to load todos
+          <br />
+          Title should not be empty
+          <br />
+          Unable to add a todo
+          <br />
+          Unable to delete a todo
+          <br />
+          Unable to update a todo */}
+    </div>
+  );
+};
